feat: Add About and Skills components
This commit adds the About and Skills components to the project. These components are responsible for rendering the About and Skills sections of the website respectively. The About component displays a section with the heading "About" and the Skills component displays a section with the heading "Skills". These components are implemented as functional components and are exported for use in other parts of the application.

The commit also includes updates to the package.json file to add new dependencies. The axios package is added with version 1.7.2 and the react-icons package is added with version 5.2.1.

Additionally, the globals.css file is modified to import fonts from Google Fonts. The font families "Caveat" and "Playwrite DE SAS" are imported with specific font weights and are set as the font families for the entire application.

Lastly, the layout.tsx file is updated to import site data from a constants file and use it to set the title and description metadata for the website.

These changes are necessary to enhance the functionality and appearance of the website.