feat: Add ScrollToTop component
This commit adds the ScrollToTop component to the project. The ScrollToTop component is responsible for displaying a button that allows users to scroll back to the top of the page with a smooth animation. It appears when the user scrolls down more than 50 pixels. This feature improves the user experience by providing an easy way to navigate back to the top of the page without manually scrolling.