style: Improve button styling in ContactForm and MaintenanceMode components
The ContactForm component and MaintenanceMode component have been updated to improve the styling of buttons. In the ContactForm component, the background color of the submit button has been changed from blue-500 to blue-600, and the hover background color has been changed from blue-600 to blue-700. In the MaintenanceMode component, the button background color has been changed from primary-700 to primary-800 on hover. These changes enhance the visual appearance and user experience of the buttons in both components.